+++ packages/netstring/test/test-netstring.js
@@ -27,7 +27,7 @@ test('read short messages', async t => {
   );
 });
 
-test('read messages divided over chunk boundaries', async t => {
+test('read a message divided over a chunk boundary', async t => {
   const r = netstringReader(
     [encoder.encode('5:hel'), encoder.encode('lo,')],
     '<unknown>',
@@ -40,15 +40,32 @@ test('read messages divided over chunk boundaries', async t => {
   );
 });
 
+test('read messages divided over chunk boundaries', async t => {
+  const r = netstringReader(
+    [
+      encoder.encode('5:hel'),
+      encoder.encode('lo,5:world,8:good '),
+      encoder.encode('bye,'),
+    ],
+    '<unknown>',
+    1,
+  );
+  const array = await read(r);
+  t.deepEqual(
+    ['hello', 'world', 'good bye'],
+    array.map(chunk => decoder.decode(chunk)),
+  );
+});
+
 function delay(ms) {
   return new Promise(resolve => {
     setTimeout(resolve, ms);
   });
 }
 
-test('round-trip short messages', async t => {
+const makeArrayWriter = () => {
   const array = [];
-  const w = netstringWriter({
+  const writer = netstringWriter({
     async next(value) {
       // Provide some back pressure to give the producer an opportunity to make
       // the mistake of overwriting the given slice.
@@ -64,10 +81,30 @@ test('round-trip short messages', async t => {
       return { done: true };
     },
   });
-  await w.next(encoder.encode(''));
-  await w.next(encoder.encode('A'));
-  await w.next(encoder.encode('hello'));
-  await w.return();
+  return { array, writer };
+};
+
+test('round-trip short messages', async t => {
+  const { array, writer } = makeArrayWriter();
+  await writer.next(encoder.encode(''));
+  await writer.next(encoder.encode('A'));
+  await writer.next(encoder.encode('hello'));
+  await writer.return();
+
+  t.deepEqual(
+    [encoder.encode(''), encoder.encode('A'), encoder.encode('hello')],
+    await read(netstringReader(array)),
+  );
+});
+
+test('concurrent writes', async t => {
+  const { array, writer } = makeArrayWriter();
+  await Promise.all([
+    writer.next(encoder.encode('')),
+    writer.next(encoder.encode('A')),
+    writer.next(encoder.encode('hello')),
+    writer.return(),
+  ]);
 
   t.deepEqual(
     [encoder.encode(''), encoder.encode('A'), encoder.encode('hello')],